Move settings from SettingsActivity to SettingsFragment
