+++ app/src/main/AndroidManifest.xml
@@ -27,9 +27,5 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".settings.ui.SettingsActivity"
-            android:screenOrientation="nosensor">
-        </activity>
     </application>
 </manifest>
